server/requirements.txt
fastapi
uvicorn
websockets
pydantic
pydantic-settings
twilio
python-dotenv
loguru
pipecat-ai
onnxruntime
aws_sdk_bedrock_runtime